+++ src/Libraries/Microsoft.Extensions.AI/ChatCompletion/OpenTelemetryChatClient.cs
@@ -37,6 +37,8 @@ public sealed partial class OpenTelemetryChatClient : DelegatingChatClient
     private readonly ActivitySource _activitySource;
     private readonly Meter _meter;
 
+    private readonly ILogger? _logger;
+
     private readonly Histogram<int> _tokenUsageHistogram;
     private readonly Histogram<double> _operationDurationHistogram;
     private readonly Histogram<double> _timeToFirstChunkHistogram;
@@ -53,12 +55,12 @@ public sealed partial class OpenTelemetryChatClient : DelegatingChatClient
     /// <param name="innerClient">The underlying <see cref="IChatClient"/>.</param>
     /// <param name="logger">The <see cref="ILogger"/> to use for emitting any logging data from the client.</param>
     /// <param name="sourceName">An optional source name that will be used on the telemetry data.</param>
-#pragma warning disable IDE0060 // Remove unused parameter; it exists for backwards compatibility and future use
     public OpenTelemetryChatClient(IChatClient innerClient, ILogger? logger = null, string? sourceName = null)
-#pragma warning restore IDE0060
         : base(innerClient)
     {
         Debug.Assert(innerClient is not null, "Should have been validated by the base ctor");
+
+        _logger = logger;
 
         if (innerClient!.GetService<ChatClientMetadata>() is ChatClientMetadata metadata)
         {
@@ -712,6 +714,11 @@ public sealed partial class OpenTelemetryChatClient : DelegatingChatClient
             _ = activity?
                 .AddTag(OpenTelemetryConsts.Error.Type, error.GetType().FullName)
                 .SetStatus(ActivityStatusCode.Error, error.Message);
+
+            if (_logger is not null)
+            {
+                OpenTelemetryLog.OperationException(_logger, error);
+            }
         }
 
         if (response is not null)